Revert "FE branch merge to main."

+++ FE/src/apis/group.ts
@@ -1,18 +1,20 @@
-import { API_URL } from '@/constants/url';
-import { GroupPostParams, GroupPutParams } from '@/types/group';
+import { GroupPostParams } from '@/types/group';
 import { getCookie } from '@/utils/cookie';
 
 export const getGroupMembers = async (mapId: number) => {
   const accessToken = getCookie('access_token');
-  if (!accessToken) window.location.href = '/login';
+  if (!accessToken) throw new Error('accessToken is undefined');
 
-  const response = await fetch(`${API_URL}/map/${mapId}/groups`, {
-    method: 'GET',
-    headers: {
-      Authorization: `Bearer ${accessToken}`,
-      'Content-Type': 'application/json',
-    },
-  });
+  const response = await fetch(
+    `${process.env.SQUAD_MAP_OAUTH_URL}/groups/${mapId}`,
+    {
+      method: 'GET',
+      headers: {
+        Authorization: `Bearer ${accessToken}`,
+        'Content-Type': 'application/json',
+      },
+    }
+  );
 
   const groupMembers = await response.json();
 
@@ -28,16 +30,19 @@ export const postGroupMember = async (
   groupPostBody: GroupPostParams
 ) => {
   const accessToken = getCookie('access_token');
-  if (!accessToken) window.location.href = '/login';
+  if (!accessToken) throw new Error('accessToken is undefined');
 
-  const response = await fetch(`${API_URL}/map/${mapId}/groups`, {
-    method: 'POST',
-    headers: {
-      Authorization: `Bearer ${accessToken}`,
-      'Content-Type': 'application/json',
-    },
-    body: JSON.stringify(groupPostBody),
-  });
+  const response = await fetch(
+    `${process.env.SQUAD_MAP_OAUTH_URL}/groups/${mapId}`,
+    {
+      method: 'POST',
+      headers: {
+        Authorization: `Bearer ${accessToken}`,
+        'Content-Type': 'application/json',
+      },
+      body: JSON.stringify(groupPostBody),
+    }
+  );
 
   const groupMember = await response.json();
 
@@ -48,56 +53,57 @@ export const postGroupMember = async (
   }
 };
 
-export const putGroupMember = async ({
-  mapId,
-  groupPutBody,
-}: {
-  mapId: number;
-  groupPutBody: GroupPutParams;
-}) => {
+export const putGroupMember = async (
+  mapId: number,
+  groupPutBody: { member_id: number; permission_level: string }
+) => {
   const accessToken = getCookie('access_token');
-  if (!accessToken) window.location.href = '/login';
+  if (!accessToken) throw new Error('accessToken is undefined');
 
-  const response = await fetch(`${API_URL}/map/${mapId}/groups`, {
-    method: 'PUT',
-    headers: {
-      Authorization: `Bearer ${accessToken}`,
-      'Content-Type': 'application/json',
-    },
-    body: JSON.stringify(groupPutBody),
-  });
+  const response = await fetch(
+    `${process.env.SQUAD_MAP_OAUTH_URL}/groups/${mapId}`,
+    {
+      method: 'PUT',
+      headers: {
+        Authorization: `Bearer ${accessToken}`,
+        'Content-Type': 'application/json',
+      },
+      body: JSON.stringify(groupPutBody),
+    }
+  );
 
-  const groupMember = await response.json();
+  const data = await response.json();
 
   try {
-    return groupMember;
+    return data;
   } catch (err) {
     throw new Error(`putGroupMember api fail err: ${err}`);
   }
 };
 
-export const deleteGroupMember = async ({
-  mapId,
-  memberId,
-}: {
-  mapId: number;
-  memberId: number;
-}) => {
+export const deleteGroupMember = async (
+  mapId: number,
+  groupDeleteBody: { selectedMemberId: number }
+) => {
   const accessToken = getCookie('access_token');
-  if (!accessToken) window.location.href = '/login';
+  if (!accessToken) throw new Error('accessToken is undefined');
 
-  const response = await fetch(`${API_URL}/map/${mapId}/groups/${memberId}`, {
-    method: 'DELETE',
-    headers: {
-      Authorization: `Bearer ${accessToken}`,
-      'Content-Type': 'application/json',
-    },
-  });
+  const response = await fetch(
+    `${process.env.SQUAD_MAP_OAUTH_URL}/groups/${mapId}`,
+    {
+      method: 'DELETE',
+      headers: {
+        Authorization: `Bearer ${accessToken}`,
+        'Content-Type': 'application/json',
+      },
+      body: JSON.stringify(groupDeleteBody),
+    }
+  );
 
-  const groupMember = await response.json();
+  const data = await response.json();
 
   try {
-    return groupMember;
+    return data;
   } catch (err) {
     throw new Error(`deleteGroupMember api fail err: ${err}`);
   }
